% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rest_b.R
\name{corr_rest_b}
\alias{corr_rest_b}
\title{corr_rest}
\usage{
corr_rest_b(phat0_e1, phat0_e2, p0_e1, p0_e2, p1_e1, p1_e2, OR1, OR2)
}
\value{
Correlation bounds based on estimated and assumed probabilities
}
\description{
computes the correlation restrictions
phat0_e1,phat0_e2: estimated probabilities
p0_e1,p0_e2,p1_e1,p1_e2: assumed probabilities (design)
OR1,OR2: expected effect sizes
}
\keyword{internal}
