% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_rest_ub.R
\name{corr_rest_ub}
\alias{corr_rest_ub}
\title{corr_rest_ub: computes the correlation restrictions unblinded case}
\usage{
corr_rest_ub(
  phat0_e1,
  phat0_e2,
  phat1_e1,
  phat1_e2,
  p0_e1,
  p0_e2,
  p1_e1,
  p1_e2,
  OR1,
  OR2
)
}
\value{
Correlation bounds based on estimated and assumed probabilities (unblinded case)
}
\description{
computes the correlation restrictions with unblinded data
phat0_e1,phat0_e2,phat1_e1,phat1_e2: estimated probabilities
p0_e1,p0_e2,p1_e1,p1_e2: assumed probabilities (design)
OR1,OR2: expected effect sizes
}
\keyword{internal}
