% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eselect.R
\name{eselect}
\alias{eselect}
\title{eselect: Endpoint selection and sample size reassessment for composite endpoints based on blinded data}
\usage{
eselect(db, p0_e1, OR1, p0_e2, OR2, criteria = "SS")
}
\arguments{
\item{db}{matrix 2x2 table (pooled sample)}

\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{OR1}{numeric parameter, Odds ratio for the endpoint 1}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{OR2}{numeric parameter, Odds ratio for the endpoint 2}

\item{criteria}{decision criteria to choose between the composite endpoint or the endpoint 1 as primary endpoint ("SS": Ratio sample sizes, "ARE": Asymptotic Relative Efficiency).}
}
\value{

}
\description{

}
\details{

}
\references{

}
