% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eselect_sim.R
\name{eselect_sim}
\alias{eselect_sim}
\title{eselect_sim}
\usage{
eselect_sim(ss_arm, p0_e1, OR1, p0_e2, OR2, p0_ce, p_init = 1, criteria = "SS")
}
\arguments{
\item{ss_arm}{numeric parameter, sample size per arm}

\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{OR1}{numeric parameter, Odds ratio for the endpoint 1}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{OR2}{numeric parameter, Odds ratio for the endpoint 2}

\item{p_init}{numeric parameter, percentage of sample size used in the interim}

\item{criteria}{decision criteria to choose between the composite endpoint or the endpoint 1 as primary endpoint ("SS": Ratio sample sizes, "ARE": Asymptotic Relative Efficiency).}
}
\value{

}
\description{

}
\details{

}
\references{

}
