% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eselectme.R
\name{eselectme}
\alias{eselectme}
\title{eselectme: Endpoint selection and sample size reassessment for multiple binary endpoints based on blinded data}
\usage{
eselectme(db, p0_e1, OR1, p0_e2, OR2, alpha = 0.05, beta = 0.2)
}
\arguments{
\item{db}{matrix 2x2 table (pooled sample)}

\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{OR1}{numeric parameter, Odds ratio for the endpoint 1}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{OR2}{numeric parameter, Odds ratio for the endpoint 2}

\item{alpha}{Type I error.}

\item{beta}{Type II error.}
}
\value{
This function returns the decision (Decision = 2, meaning the chosen endpoint is the multiple endpoints approach; Decision = 1, meaning the chosen endpoint is the composite endpoint; and Decision = 0, meaning the chosen endpoint is the relevant endpoint) and the sample size according to the decision.
}
\description{
Endpoint selection and sample size reassessment for multiple binary endpoints based on blinded data. The composite endpoint is assumed to be a binary endpoint formed by a combination of two events (E1 and E2). We assume that the endpoint 1 is more relevant for the clinical question than endpoint 2. This function selects between the composite endpoint (CE), the relevant endpoint (RE), and the multiple binary endpoints (CE and RE) as the primary endpoint(s) of the study and recalculate the sample size accordingly. The decision criteria to decide between the composite endpoint or the relevant endpoint is the ratio of the corresponding sample sizes ("SS").
The algorithm of the function is the following: First, the probabilities of the composite components in the control group and the correlation between them is estimated based on blinded data. Second, using the estimated probabilities and the estimated correlation, the decision criteria is computed and the primary endpoint is selected.  Finally, the sample size is recalculated according to the decision.
}
\details{

}
\references{

}
