% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_ub.R
\name{estimation_ub}
\alias{estimation_ub}
\title{Unblinded estimation of the correlation}
\usage{
estimation_ub(samplesize, p0_e1, p1_e1, OR1, p0_e2, p1_e2, OR2, p0_ce, p1_ce)
}
\arguments{
\item{samplesize}{numeric parameter, sample size per arm}

\item{p0_e1}{numeric parameter, probability of occurrence E1 in the control group}

\item{p1_e1}{numeric parameter, probability of occurrence E1 in the treatment group}

\item{OR1}{numeric parameter, Odds ratio for the endpoint 1 (design)}

\item{p0_e2}{numeric parameter, probability of occurrence E2 in the control group}

\item{p1_e2}{numeric parameter, probability of occurrence E2 in the treatment group}

\item{OR2}{numeric parameter, Odds ratio for the endpoint 2 (design)}

\item{p0_ce}{numeric parameter, probability of occurrence composite endpoint in the control group}

\item{p1_ce}{numeric parameter, probability of occurrence composite endpoint in the treatment group}
}
\value{
This function returns the estimated correlation based on unblinded data and the truncated correlation within the possible margins.
}
\description{
This function estimates the correlation between two binary endpoints based on unblinded data.
}
\keyword{internal}
