% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_OR.R
\name{f_OR}
\alias{f_OR}
\title{f_OR}
\usage{
f_OR(samplesize, p0, p1)
}
\value{
Two-sample test statistics using unpooled variance estimator
}
\description{
simulates binary outcomes and computes the statistic
}
\keyword{internal}
