% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{f_sim}
\alias{f_sim}
\title{f_sim: A function rather aimed at developers}
\usage{
f_sim(samplesize, p_e1, p_e2, p_ce)
}
\description{
simulation binary data two outcomes
returns 2x2 table
s1+s2: num X1
s1+s3: num X2
}
\keyword{internal}
