% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{fun_p0}
\alias{fun_p0}
\title{fun_p0}
\usage{
fun_p0(p, l)
}
\description{
computes the probability under the control group
based on the pooled probability and the odds ratio
p: pooled probability
l: odds ratio
}
\keyword{internal}
