% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_function.R
\name{OR_function}
\alias{OR_function}
\title{OR_function}
\usage{
OR_function(p0, p1)
}
\value{
Odds ratio calculation based on marginal probabilities
}
\description{
computes the odds ratio
}
\keyword{internal}
