% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize_OR.R
\name{samplesize_OR}
\alias{samplesize_OR}
\title{samplesize_OR}
\usage{
samplesize_OR(p0, OR, alpha = 0.05, beta = 0.2, Unpooled = "Unpooled Variance")
}
\value{
Sample size for odds ratios
}
\description{
Sample size calculations in terms of odds ratio
}
\keyword{internal}
