% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_me.R
\name{test_me}
\alias{test_me}
\title{test_me}
\usage{
test_me(OR1, p0_e1, OR2, p0_e2, n)
}
\value{
Two-sample test statistics for endpoint 1 and 2
}
\description{
computes the statistical tests (OR) for two endpoints
}
\keyword{internal}
