% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_sim.R
\name{f_sim}
\alias{f_sim}
\title{Simulation 2x2 table binary endpoints}
\usage{
f_sim(samplesize, p_e1, p_e2, p_ce)
}
\arguments{
\item{samplesize}{sample size simulated sample}

\item{p_e1}{numeric parameter, probability of occurrence E1}

\item{p_e2}{numeric parameter, probability of occurrence E2}

\item{p_ce}{numeric parameter, probability of occurrence composite endpoint}
}
\value{
Simlated data

the function returns 2x2 table
s1+s2: num X1
s1+s3: num X2
}
\description{
simulation binary data two outcomes
}
\keyword{internal}
